Add Kernel for lenvo X1 carbon
Should put this in a readme as well but this kernel supports NVME
out of the box as well as having built in support for running Docker.
It has limited filesystem support with only ext3/ext4 build into the
kernel. This is capable of booting the lenovo x1 carbon without an
initramfs if you so choose to do.